Added header guards for Channel
