+++ src/resonanceReconstruction/rmatrix/Channel.hpp
@@ -1,3 +1,30 @@
+#ifndef NJOY_R2_RMATRIX_CHANNEL
+#define NJOY_R2_RMATRIX_CHANNEL
+
+// system includes
+
+// other includes
+#include "elementary/ReactionID.hpp"
+#include "resonanceReconstruction/quantities.hpp"
+#include "resonanceReconstruction/rmatrix/ChannelQuantumNumbers.hpp"
+#include "resonanceReconstruction/rmatrix/ParticlePair.hpp"
+#include "resonanceReconstruction/rmatrix/ChannelRadii.hpp"
+#include "resonanceReconstruction/rmatrix/calculatePenetrability.hpp"
+#include "resonanceReconstruction/rmatrix/calculateShiftFactor.hpp"
+#include "resonanceReconstruction/rmatrix/calculatePhaseShift.hpp"
+#include "resonanceReconstruction/rmatrix/calculateCoulombPhaseShift.hpp"
+
+namespace njoy {
+namespace resonanceReconstruction {
+namespace rmatrix {
+
+// identifiers
+using ReactionID = elementary::ReactionID;
+using ChannelID = std::string;
+
+// boundary condition
+using BoundaryCondition = double;
+
 /**
  *  @class
  *  @brief A reaction channel
@@ -106,3 +133,9 @@ public:
   #include "resonanceReconstruction/rmatrix/Channel/src/phaseShift.hpp"
   #include "resonanceReconstruction/rmatrix/Channel/src/coulombPhaseShift.hpp"
 };
+
+} // rmatrix namespace
+} // resonanceReconstruction namespace
+} // njoy namespace
+
+#endif
