+++ src/resonanceReconstruction/rmatrix/ChannelTypes.hpp
@@ -1,0 +1,22 @@
+#ifndef NJOY_R2_RMATRIX_CHANNELTYPES
+#define NJOY_R2_RMATRIX_CHANNELTYPES
+
+// system includes
+
+// other includes
+
+namespace njoy {
+namespace resonanceReconstruction {
+namespace rmatrix {
+
+  // particle and channel types
+  struct Neutron {};
+  struct Photon {};
+  struct ChargedParticle {};
+  struct Fission {};
+
+} // rmatrix namespace
+} // resonanceReconstruction namespace
+} // njoy namespace
+
+#endif
