+++ src/resonanceReconstruction/rmatrix/calculateShiftFactor.hpp
@@ -1,3 +1,18 @@
+#ifndef NJOY_R2_RMATRIX_CALCULATES
+#define NJOY_R2_RMATRIX_CALCULATES
+
+// system includes
+#include <complex>
+
+// other includes
+#include "utility/horner.hpp"
+#include "resonanceReconstruction/rmatrix/src/coh3-coulomb.hpp"
+#include "resonanceReconstruction/rmatrix/ChannelTypes.hpp"
+
+namespace njoy {
+namespace resonanceReconstruction {
+namespace rmatrix {
+
 /**
  *  @brief Default value for the shift factor
  *
@@ -29,22 +44,34 @@ double calculateShiftFactor< Neutron >( const unsigned int l,
 
       constexpr std::array< double, 2 > numerator = {{ 18., 3. }};
       constexpr std::array< double, 3 > denominator = {{ 9., 3., 1. }};
-      return -horner( std::rbegin( numerator ), std::rend( numerator ), squared )
-             / horner( std::rbegin( denominator ), std::rend( denominator ), squared );
+      return -utility::horner( std::rbegin( numerator ),
+                               std::rend( numerator ),
+                               squared )
+             / utility::horner( std::rbegin( denominator ),
+                                std::rend( denominator ),
+                                squared );
     }
     case 3 : {
 
       constexpr std::array< double, 3 > numerator = {{ 675., 90., 6. }};
       constexpr std::array< double, 4 > denominator = {{ 225., 45., 6., 1. }};
-      return -horner( std::rbegin( numerator ), std::rend( numerator ), squared )
-             / horner( std::rbegin( denominator ), std::rend( denominator ), squared );
+      return -utility::horner( std::rbegin( numerator ),
+                               std::rend( numerator ),
+                               squared )
+             / utility::horner( std::rbegin( denominator ),
+                                std::rend( denominator ),
+                                squared );
     }
     case 4 : {
 
       constexpr std::array< double, 4 > numerator = {{ 44100., 4725., 270., 10. }};
       constexpr std::array< double, 5 > denominator = {{ 11025., 1575., 135., 10., 1. }};
-      return -horner( std::rbegin( numerator ), std::rend( numerator ), squared )
-             / horner( std::rbegin( denominator ), std::rend( denominator ), squared );
+      return -utility::horner( std::rbegin( numerator ),
+                               std::rend( numerator ),
+                               squared )
+             / utility::horner( std::rbegin( denominator ),
+                                std::rend( denominator ),
+                                squared );
     }
     default : throw std::exception();
   }
@@ -76,3 +103,9 @@ double calculateShiftFactor< ChargedParticle >( const unsigned int l,
            ? 0.
            : ratio / ( F * F + G * G ) * ( F * dF + G * dG );
 }
+
+} // rmatrix namespace
+} // resonanceReconstruction namespace
+} // njoy namespace
+
+#endif
