+++ src/resonanceReconstruction/rmatrix/src/coh3-coulomb.hpp
@@ -1,3 +1,6 @@
+#ifndef NJOY_R2_RMATRIX_COH3_COULOMB
+#define NJOY_R2_RMATRIX_COH3_COULOMB
+
 //const double HBAR       =  6.582119514e-22; /* Planck's constant/2pi [MeV sec]*/
 //const double VLIGHT     =  2.99792458e+23 ; /* light velocty [fm/sec]         */
 //const double AMUNIT     =  931.4943335    ; /* MeV = 1amu                     */
@@ -36,3 +39,5 @@ void coulombWaveFunctions( int l, double ratio, double eta,
   delete [] Cou0;
   delete [] Cou1;
 }
+
+#endif
